refactor: 팀 내 팀원들 시간표 목록 조회  기능을 팀 상세 조회로 편입, DTO 최적화
