+++ src/main/java/com/jandi/band_backend/team/controller/TeamTimetableController.java
@@ -2,7 +2,9 @@ package com.jandi.band_backend.team.controller;
 
 import com.jandi.band_backend.global.CommonResponse;
 import com.jandi.band_backend.security.CustomUserDetails;
-import com.jandi.band_backend.team.dto.*;
+import com.jandi.band_backend.team.dto.ScheduleSuggestionRespDTO;
+import com.jandi.band_backend.team.dto.TimetableReqDTO;
+import com.jandi.band_backend.team.dto.TimetableRespDTO;
 import com.jandi.band_backend.team.service.TeamTimetableService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -33,17 +35,6 @@ public class TeamTimetableController {
                 .body(CommonResponse.success("스케줄 조율 모드가 시작되었습니다", result));
     }
 
-    @Operation(summary = "팀내 팀원들 시간표 목록 조회")
-    @GetMapping("/{teamId}/timetables")
-    public ResponseEntity<CommonResponse<TeamTimetablesRespDTO>> getTeamTimetables(
-            @PathVariable Integer teamId,
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        Integer currentUserId = userDetails.getUserId();
-        TeamTimetablesRespDTO result = teamTimetableService.getTeamTimetables(teamId, currentUserId);
-        return ResponseEntity.ok(CommonResponse.success("팀 시간표 목록 조회 성공", result));
-    }
-
     @Operation(summary = "팀내 내 시간표 입력")
     @PostMapping("/{teamId}/members/me/timetable")
     public ResponseEntity<CommonResponse<TimetableRespDTO>> submitMyTimetable(
@@ -55,6 +46,4 @@ public class TeamTimetableController {
         TimetableRespDTO result = teamTimetableService.submitMyTimetable(teamId, reqDTO, currentUserId);
         return ResponseEntity.ok(CommonResponse.success("팀 시간표 입력 성공", result));
     }
-
-
 }
