+++ src/main/java/com/jandi/band_backend/team/service/TeamService.java
@@ -1,5 +1,8 @@
 package com.jandi.band_backend.team.service;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jandi.band_backend.club.entity.Club;
 import com.jandi.band_backend.club.entity.ClubMember;
 import com.jandi.band_backend.club.repository.ClubMemberRepository;
@@ -18,6 +21,7 @@ import com.jandi.band_backend.team.repository.TeamRepository;
 import com.jandi.band_backend.user.entity.Users;
 import com.jandi.band_backend.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -25,8 +29,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -37,6 +43,7 @@ public class TeamService {
     private final ClubRepository clubRepository;
     private final ClubMemberRepository clubMemberRepository;
     private final UserRepository userRepository;
+    private final ObjectMapper objectMapper;
 
     /**
      * 곡 팀 생성
@@ -60,6 +67,8 @@ public class TeamService {
         team.setClub(club);
         team.setName(teamReqDTO.getName());
         team.setCreator(currentUser);
+        // 팀 생성과 동시에 suggestedScheduleAt에 현재 시간 설정
+        team.setSuggestedScheduleAt(LocalDateTime.now());
 
         Team savedTeam = teamRepository.save(team);
 
@@ -97,7 +106,7 @@ public class TeamService {
     }
 
     /**
-     * 팀 상세 조회
+     * 팀 상세 조회 (시간표 정보 포함)
      */
     public TeamDetailRespDTO getTeamDetail(Integer teamId, Integer currentUserId) {
         // 팀 존재 확인
@@ -111,7 +120,7 @@ public class TeamService {
         // 팀 멤버 목록 조회
         List<TeamMember> teamMembers = teamMemberRepository.findByTeamId(teamId);
 
-        return createTeamDetailRespDTO(team, teamMembers);
+        return createTeamDetailRespDTOWithTimetable(team, teamMembers);
     }
 
     /**
@@ -155,6 +164,126 @@ public class TeamService {
     }
 
     /**
+     * TeamRespDTO 생성
+     */
+    private TeamRespDTO createTeamRespDTO(Team team, Integer memberCount) {
+        return TeamRespDTO.builder()
+                .id(team.getId())
+                .name(team.getName())
+                .clubId(team.getClub().getId())
+                .clubName(team.getClub().getName())
+                .creatorId(team.getCreator().getId())
+                .creatorName(team.getCreator().getNickname())
+                .memberCount(memberCount)
+                .createdAt(team.getCreatedAt())
+                .build();
+    }
+
+    /**
+     * TeamDetailRespDTO 생성
+     */
+    private TeamDetailRespDTO createTeamDetailRespDTO(Team team, List<TeamMember> teamMembers) {
+        return TeamDetailRespDTO.builder()
+                .id(team.getId())
+                .name(team.getName())
+                .clubId(team.getClub().getId())
+                .clubName(team.getClub().getName())
+                .creatorId(team.getCreator().getId())
+                .creatorName(team.getCreator().getNickname())
+                .members(teamMembers.stream()
+                        .map(this::createMemberInfoDTO)
+                        .collect(Collectors.toList()))
+                .suggestedScheduleAt(team.getSuggestedScheduleAt())
+                .createdAt(team.getCreatedAt())
+                .updatedAt(team.getUpdatedAt())
+                .build();
+    }
+
+        /**
+     * TeamDetailRespDTO 생성 (시간표 정보 포함)
+     */
+    private TeamDetailRespDTO createTeamDetailRespDTOWithTimetable(Team team, List<TeamMember> teamMembers) {
+        // 시간표 정보 설정 (새로 생성되는 팀은 항상 suggestedScheduleAt이 존재)
+
+        // 제출 현황 계산
+        int submittedCount = 0;
+        if (team.getSuggestedScheduleAt() != null) {
+            for (TeamMember teamMember : teamMembers) {
+                boolean isSubmitted = teamMember.getUpdatedTimetableAt() != null &&
+                        teamMember.getUpdatedTimetableAt().isAfter(team.getSuggestedScheduleAt());
+                if (isSubmitted) {
+                    submittedCount++;
+                }
+            }
+        }
+
+        TeamDetailRespDTO.SubmissionProgressDTO submissionProgress = TeamDetailRespDTO.SubmissionProgressDTO.builder()
+                .submittedMember(submittedCount)
+                .totalMember(teamMembers.size())
+                .build();
+
+        return TeamDetailRespDTO.builder()
+                .id(team.getId())
+                .name(team.getName())
+                .clubId(team.getClub().getId())
+                .clubName(team.getClub().getName())
+                .creatorId(team.getCreator().getId())
+                .creatorName(team.getCreator().getNickname())
+                .members(teamMembers.stream()
+                        .map(teamMember -> createMemberInfoDTOWithTimetable(teamMember, team.getSuggestedScheduleAt()))
+                        .collect(Collectors.toList()))
+                .suggestedScheduleAt(team.getSuggestedScheduleAt())
+                .submissionProgress(submissionProgress)
+                .createdAt(team.getCreatedAt())
+                .updatedAt(team.getUpdatedAt())
+                .build();
+    }
+
+    /**
+     * MemberInfoDTO 생성
+     */
+    private TeamDetailRespDTO.MemberInfoDTO createMemberInfoDTO(TeamMember teamMember) {
+        return TeamDetailRespDTO.MemberInfoDTO.builder()
+                .userId(teamMember.getUser().getId())
+                .name(teamMember.getUser().getNickname())
+                .position(teamMember.getUser().getPosition() != null ?
+                        teamMember.getUser().getPosition().name() : null)
+                .build();
+    }
+
+    /**
+     * MemberInfoDTO 생성 (시간표 정보 포함)
+     */
+    private TeamDetailRespDTO.MemberInfoDTO createMemberInfoDTOWithTimetable(TeamMember teamMember, LocalDateTime suggestedScheduleAt) {
+        // 시간표 데이터 파싱
+        Map<String, List<String>> timetableData = null;
+        if (teamMember.getTimetableData() != null) {
+            try {
+                timetableData = objectMapper.readValue(teamMember.getTimetableData(),
+                        new TypeReference<Map<String, List<String>>>() {});
+            } catch (JsonProcessingException e) {
+                log.error("시간표 데이터 파싱 오류: {}", e.getMessage());
+            }
+        }
+
+        // 제출 여부 확인
+        boolean isSubmitted = false;
+        if (suggestedScheduleAt != null && teamMember.getUpdatedTimetableAt() != null) {
+            isSubmitted = teamMember.getUpdatedTimetableAt().isAfter(suggestedScheduleAt);
+        }
+
+        return TeamDetailRespDTO.MemberInfoDTO.builder()
+                .userId(teamMember.getUser().getId())
+                .name(teamMember.getUser().getNickname())
+                .position(teamMember.getUser().getPosition() != null ?
+                        teamMember.getUser().getPosition().name() : null)
+                .timetableUpdatedAt(teamMember.getUpdatedTimetableAt())
+                .isSubmitted(isSubmitted)
+                .timetableData(timetableData)
+                .build();
+    }
+
+    /**
      * 팀 수정/삭제 권한 확인 (팀 생성자 또는 동아리 대표자)
      */
     private void validateTeamModificationPermission(Team team, Integer currentUserId) {
@@ -171,69 +300,5 @@ public class TeamService {
         if (!isCreator && !isRepresentative) {
             throw new UnauthorizedClubAccessException("팀 생성자 또는 동아리 대표자만 팀 정보를 수정할 수 있습니다.");
         }
-    }
-
-    /**
-     * TeamRespDTO 생성
-     */
-    private TeamRespDTO createTeamRespDTO(Team team, Integer memberCount) {
-        return TeamRespDTO.builder()
-                .id(team.getId())
-                .name(team.getName())
-                .creatorId(team.getCreator().getId())
-                .creatorName(team.getCreator().getNickname())
-                .memberCount(memberCount)
-                .createdAt(team.getCreatedAt())
-                .build();
-    }
-
-    /**
-     * TeamDetailRespDTO 생성
-     */
-    private TeamDetailRespDTO createTeamDetailRespDTO(Team team, List<TeamMember> teamMembers) {
-        return TeamDetailRespDTO.builder()
-                .id(team.getId())
-                .name(team.getName())
-                .club(createClubInfoDTO(team.getClub()))
-                .creator(createCreatorInfoDTO(team.getCreator()))
-                .members(teamMembers.stream()
-                        .map(this::createMemberInfoDTO)
-                        .collect(Collectors.toList()))
-                .memberCount(teamMembers.size())
-                .createdAt(team.getCreatedAt())
-                .updatedAt(team.getUpdatedAt())
-                .build();
-    }
-
-    /**
-     * ClubInfoDTO 생성
-     */
-    private TeamDetailRespDTO.ClubInfoDTO createClubInfoDTO(Club club) {
-        return TeamDetailRespDTO.ClubInfoDTO.builder()
-                .clubId(club.getId())
-                .name(club.getName())
-                .build();
-    }
-
-    /**
-     * CreatorInfoDTO 생성
-     */
-    private TeamDetailRespDTO.CreatorInfoDTO createCreatorInfoDTO(Users user) {
-        return TeamDetailRespDTO.CreatorInfoDTO.builder()
-                .userId(user.getId())
-                .name(user.getNickname())
-                .build();
-    }
-
-    /**
-     * MemberInfoDTO 생성
-     */
-    private TeamDetailRespDTO.MemberInfoDTO createMemberInfoDTO(TeamMember teamMember) {
-        return TeamDetailRespDTO.MemberInfoDTO.builder()
-                .userId(teamMember.getUser().getId())
-                .name(teamMember.getUser().getNickname())
-                .position(teamMember.getUser().getPosition() != null ?
-                        teamMember.getUser().getPosition().name() : null)
-                .build();
     }
 }
